Feature dammy data

+++ WOB/Models/Player.cs
@@ -10,9 +10,9 @@ namespace WOB.Models
         public string FirstName { get; set; }
 
         public string LastName { get; set; }
-        public float Height { get; set; }
+        public double Height { get; set; }
 
-        public float Weight { get; set; }
+        public double Weight { get; set; }
         [ForeignKey("Status")]
         public int StatusId { get; set; }
